requirements.txt
certifi==2017.11.5
chardet==3.0.4
idna==2.6
requests==2.18.4
urllib3==1.22
geolocation-python==0.2.2
pyTelegramBotAPI==3.5.1
python-bandsintown==0.2.0
peewee==2.10.2
